Problem with early termination when looking in visited nodes

+++ pathfinder/searchspace.h
@@ -23,8 +23,8 @@ public:
 	searchSpace(const int nMapWidth, const int nMapHeight, unsigned int* const * pMap) :
 		m_nMapWidth(nMapWidth), m_nMapHeight(nMapHeight), m_pMap(pMap)
 	{
-		m_vNodeVector.reserve(nMapWidth * nMapHeight / 9);
-		m_visitedNodes.reserve(nMapWidth * nMapHeight / 9);
+		m_vNodeVector.reserve(int(nMapWidth * nMapHeight / 9));
+		m_visitedNodes.reserve(int(nMapWidth * nMapHeight / 9));
 	}
 	bool addNeighbouringNodes();
 	bool insertInitialNodes(const vec2& nStartPos, const vec2& nGoalPos);
